# 49_Se7enSquare-main/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/communities": {
      "get": {
        "operationId": "CommunitiesController_findAll",
        "summary": "List all communities",
        "description": "Returns all communities including enriched UI fields (icon, category, slug, memberCount, onlineCount).",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of community records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "communities"
        ]
      },
      "post": {
        "operationId": "CommunitiesController_create",
        "summary": "Create a community",
        "description": "Creates a new community. Any authenticated role may create a community.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Community creation payload (name, description, ownerId, tags)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created community",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "communities"
        ]
      }
    },
    "/communities/{id}": {
      "get": {
        "operationId": "CommunitiesController_findOne",
        "summary": "Get a single community by ID",
        "description": "Fetches one community by its numeric ID, including all display fields.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric community ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching community record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "Community not found"
          }
        },
        "tags": [
          "communities"
        ]
      },
      "patch": {
        "operationId": "CommunitiesController_update",
        "summary": "Update a community",
        "description": "Partially updates a community. All fields are optional.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric community ID to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update (all optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated community record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "Community not found"
          }
        },
        "tags": [
          "communities"
        ]
      },
      "delete": {
        "operationId": "CommunitiesController_remove",
        "summary": "Delete a community",
        "description": "Permanently deletes a community and cascades to related memberships and events. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric community ID to delete",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Community 3 deleted"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin can delete communities"
          },
          "404": {
            "description": "Community not found"
          }
        },
        "tags": [
          "communities"
        ]
      }
    },
...
  },
  "info": {
    "title": "Se7enSquare Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | moderator | user | gamer"
      }
    },
    "schemas": {
...
      "CommunityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique community ID"
          },
          "name": {
            "type": "string",
            "example": "FPS Arena",
            "description": "Community display name"
          },
          "description": {
            "type": "string",
            "example": "Competitive FPS players and tournaments",
            "description": "Community description (10–200 chars)"
          },
          "ownerId": {
            "type": "number",
            "example": 3,
            "description": "User ID of the community owner"
          },
          "tags": {
            "example": [
              "fps",
              "esports"
            ],
            "description": "Topic tags (1–5 tags)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "icon": {
            "type": "string",
            "example": "⚡",
            "description": "Emoji icon shown in the UI"
          },
          "category": {
            "type": "string",
            "example": "Gaming",
            "description": "Category label shown in discovery page"
          },
          "slug": {
            "type": "string",
            "example": "fps-arena",
            "description": "URL-friendly slug for routing"
          },
          "memberCount": {
            "type": "number",
            "example": 12400,
            "description": "Total member count"
          },
          "onlineCount": {
            "type": "number",
            "example": 842,
            "description": "Members currently online"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "ownerId",
          "tags"
        ]
      },
      "CreateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "MOBA Masters"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "example": "Community for strategy and ranked discussions"
          },
          "ownerId": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "example": [
              "moba",
              "ranked"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "description",
          "ownerId",
          "tags"
        ]
      },
      "UpdateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "MOBA Masters"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "example": "Community for strategy and ranked discussions"
          },
          "ownerId": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "example": [
              "moba",
              "ranked"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}